chore: restructured package structure
