+++ internal/values/timespan.go
